backend/requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.1
DateTime==5.5
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
MarkupSafe==3.0.2
psycopg2==2.9.10
pycparser==2.23
python-dotenv==1.1.1
pytz==2025.2
setuptools==80.9.0
Werkzeug==3.1.3
zope.interface==8.0
